Allied players affected by pulsing.
Accepted by balancing on 02/01/2022^[1]

"This is a balancing change and should only be accepted when the balancing team have come to a decision concerning it.

Previously, gravity pulsing only affected unallied players. In this commit, it affects everyone. (Because why wouldn't it?)
I suppose you could argue that these pulses are actually precise beams which target enemy ships?

This commit encourages a fleet composition of smaller faster ships which pursue a larger target as opposed to just reducing the larger target's cruis..."

[1] Archived Balancing 1: https://discord.com/channels/916891217596395631/921817419348066344/927227520338755655