changed one character, maybe fixed hyperspace releasing
The check for hyperspace dimensions in the tryRelease function checks for capital-H "Hyperspace" in the dimension name. To my knowledge, all hyperspace dimensions on Horizon's End use a lowercase-h (the checks in SimpleProjectile.kt and StarshipMovement.kt also use lowercase-h). Assuming this is the case, this check would return false even in a dimension such as Asteri_hyperspace; now, it will return true.